+++ e2e/test/jest-react/stryker.conf.js
@@ -4,6 +4,7 @@ module.exports = function (config) {
     mutator: "javascript",
     transpilers: [],
     reporters: ["event-recorder"],
+    tempDirName: "stryker-tmp",
     coverageAnalysis: "off",
     timeoutMS: 60000, // High timeout to survive high build server load. Mutants created here should never timeout
     mutate: ["src/{Alert,Badge,Breadcrumb}.js"],