+++ packages/api/src/plugin/Plugins.ts
@@ -7,6 +7,8 @@ import { TestFramework } from '../../test_framework';
 import { TestRunner } from '../../test_runner';
 import { Transpiler } from '../../transpile';
 import { OptionsEditor } from '../core/OptionsEditor';
+import { TestRunner2 } from '../../test_runner2';
+import { Checker } from '../../check';
 
 import { PluginContexts } from './Contexts';
 import { PluginKind } from './PluginKind';
@@ -89,7 +91,9 @@ export interface PluginInterfaces {
   [PluginKind.Reporter]: Reporter;
   [PluginKind.TestFramework]: TestFramework;
   [PluginKind.TestRunner]: TestRunner;
+  [PluginKind.TestRunner2]: TestRunner2;
   [PluginKind.Transpiler]: Transpiler;
+  [PluginKind.Checker]: Checker;
 }
 
 /**