+++ packages/util/src/index.ts
@@ -7,3 +7,5 @@ export * from './stringUtils';
 export * from './noopLogger';
 export * from './notEmpty';
 export * from './flatMap';
+export * from './I';
+export * from './Task';